aws s3 added

+++ src/main/java/com/example/startapp/entity/User.java
@@ -28,7 +28,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private Long userId;
 
     @NotBlank(message = "The name field can't be blank")
     private String name;
@@ -78,14 +78,6 @@ public class User implements UserDetails {
     private LocalDateTime updatedDate;
 
     private boolean emailVerified;
-
-//    @NotBlank(message = "The phone prefix field can't be blank")
-//    private String phonePrefix;
-//
-//    @NotBlank(message = "The phone number field can't be blank")
-//    @Pattern(regexp = "^\\d{7}$",
-//            message = "Phone number must be 7 digits")
-//    private String phoneNumber;
 
     @OneToMany(mappedBy = "user",cascade = CascadeType.ALL)
     List<Ad> ads;
